Fancier delay FX with an LFO. Some weird effects can be produced, but beauty and style may be lacking.

+++ TLfo.h
@@ -33,5 +33,5 @@ public:
 private:
     float Hz;
     TFraction Value;
-    float Scanpos;
+    double Scanpos;
 };
